COD-31 feature(djangoProject) finish with skeleton of site
