+++ djangoProject/botArena/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class PollsConfig(AppConfig):
-    name = 'polls'
+    name = 'botArena'
