+++ djangoProject/botArena/templates/botArena/game.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <h1>{{ game.name }}</h1>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<strong>Страница для игры</strong>
+ <li>{{ game}}</li>
+<li>Я буду передавать сюда объект game, параметры смотри в models.Game</li>
+</body>
+</html>
