+++ djangoProject/botArena/templates/botArena/home.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Домашняя страница</title>
+    <strong>Тут Егор сделает домашнюю страницу</strong>
+    <li><a href="{% url 'botArena:logout'  %}">{{ "Выйти из аккаунта"}}</a></li>
+
+</head>
+<body>
+<strong>Сюда я передаю game_list, что есть QuerySet из игр.</strong>
+{% if game_list %}
+    <ul>
+    {% for game in game_list %}
+        <li><a href="{% url 'botArena:game' game.name %}">{{ game.name }}</a> Лучший игрок {{ game.leader }} Лучший счет {{ game.leader_score }}</li>
+    {% endfor %}
+    </ul>
+{% else %}
+    <p>No games are available.</p>
+{% endif %}
+</body>
+</html>
