+++ djangoProject/botArena/templates/botArena/logout.html
