+++ djangoProject/botArena/templates/botArena/results.html
@@ -6,4 +6,4 @@
 {% endfor %}
 </ul>
 
-<a href="{% url 'polls:detail' question.id %}">Vote again?</a>
+<a href="{% url 'botArena:detail' question.id %}">Vote again?</a>
