--- djangoProject/polls/templates/polls/game.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<strong>Страница для игры  (нужно будет плевать сюда весь объект game)</strong>
-</body>
-</html>
