--- djangoProject/polls/templates/polls/home.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Домашняя страница</title>
-    <strong>Тут Егор сделает домашнюю страницу</strong>
-    <li><a href="{% url 'polls:logout'  %}">{{ "Выйти из аккаунта"}}</a></li>
-
-</head>
-<body>
-{% if game_list %}
-    <ul>
-    {% for game in game_list %}
-        <li><a href="{% url 'polls:game' game.name %}">{{ game.name }}</a></li>
-    {% endfor %}
-    </ul>
-{% else %}
-    <p>No polls are available.</p>
-{% endif %}
-</body>
-</html>
