--- djangoProject/polls/urls.py
@@ -1,20 +1,0 @@
-from django.urls import path
-
-from . import views
-
-app_name = 'polls'
-urlpatterns = [
-    # ex: /polls/
-    path('', views.index, name='index'),
-    # ex: /polls/5/
-    path('<int:question_id>/', views.detail, name='detail'),
-    # ex: /polls/5/results/
-    path('<int:question_id>/results/', views.results, name='results'),
-    # ex: /polls/5/vote/
-    path('<int:question_id>/vote/', views.vote, name='vote'),
-    path('logging_test/', views.logging_test, name="login"),
-    path('home/', views.home, name='home'),
-    path('logout/', views.logout_view, name='logout'),
-    path('registration/', views.registration, name ='registration'),
-    path('game/<str:name>',views.game, name = 'game' )
-]
